# api/openapi.yaml
openapi: 3.1.0
info:
  title: IOE Backend API
  version: 0.1.0
  description: >
    Single-tenant LMS backend. Authentication exchanges a Google ID token for an
    access token (Bearer, 15 minutes) and a rotating refresh token held in an
    HttpOnly cookie scoped to /v1/auth.
servers:
  - url: https://api.example.com
paths:
...
  /v1/courses/{courseID}/lectures/{lectureID}/quizzes:
    get:
      summary: List a lecture's quizzes with answer keys
      description: >
        Anyone who may read the lecture's content: a course manager, anyone on a free-preview
        lecture of a published course, or an actively enrolled student. The client grades
        locally, so answer keys are included. Ordered by position, then ID.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/LectureID"
        - $ref: "#/components/parameters/VersionQuery"
      responses:
        "200":
          description: Quizzes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Quiz" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      summary: Create a quiz in a lecture
      description: The course owner or a root admin, on a course that is not archived. Question and option IDs are generated.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/LectureID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaveQuizRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/courses/{courseID}/lectures/{lectureID}/quizzes/{quizID}:
    put:
      summary: Replace a quiz's position and questions
      description: >
        The course owner or a root admin. Supplied question and option IDs must already belong
        to this quiz; omitted IDs are generated.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/CourseID"
        - $ref: "#/components/parameters/LectureID"
        - $ref: "#/components/parameters/QuizID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaveQuizRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/quizzes/{quizID}:
    delete:
      summary: Delete a quiz and its attempts
      description: The course owner or a root admin. Lecture blocks referencing the quiz are kept.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/QuizID"
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /v1/quizzes/{quizID}/attempts:
    post:
      summary: Record a quiz attempt
      description: >
        Only the user themselves (`user_id` must be the caller), with an active enrollment.
        Answers are stored as submitted; the client grades. Retrying with the same
        Idempotency-Key returns the first attempt's ID.
      security:
        - bearer: []
      parameters:
        - $ref: "#/components/parameters/QuizID"
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecordQuizAttemptRequest" }
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizAttemptResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Secure-ioe_refresh
  parameters:
    VersionQuery:
      name: version
      in: query
      required: false
      schema: { type: integer, minimum: 1 }
      description: Published version to read; managers only. Omit for the working copy (managers) or the live version (everyone else).
...
    CourseID:
      name: courseID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
    LectureID:
      name: lectureID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
    QuizID:
      name: quizID
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ID" }
...
  responses:
...
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    MethodNotAllowed:
      description: Method Not Allowed
      headers:
        Allow:
          description: Allowed HTTP methods for the target resource
          required: true
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InternalError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ID:
      type: string
      pattern: '^[1-9][0-9]{0,18}$'
      description: Snowflake ID rendered as a decimal string.
...
    Problem:
      type: object
      required: [type, code, title, status]
      properties:
        type: { $ref: "#/components/schemas/ProblemType" }
        code:
          allOf: [{ $ref: "#/components/schemas/ProblemType" }]
          description: Same value as type, for clients that read a code member.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, description: Request ID }
        details:
          type: object
          description: >
            Present on exam_not_open (opens_at), exam_closed (closes_at), reveal_not_yet
            (reveal_at), and edit_* codes (question_id, option_id when known).
          properties:
            opens_at: { type: string, format: date-time }
            closes_at: { type: string, format: date-time }
            reveal_at: { type: string, format: date-time }
            question_id: { $ref: "#/components/schemas/ID" }
            option_id: { $ref: "#/components/schemas/ID" }
    ProblemType:
      type: string
      enum:
        - invalid_request
        - invalid_token
        - refresh_reuse_detected
        - email_unverified
        - origin_not_allowed
        - not_found
        - method_not_allowed
        - payload_too_large
        - unsupported_media_type
        - rate_limited
        - internal
        - forbidden
        - enrollment_required
        - concurrent_modification
        - course_not_editable
        - invalid_transition
        - revision_conflict
        - lecture_content_revision_required
        - patch_too_large
        - block_set_mismatch
        - order_delete_overlap
        - duplicate_client_block_id
        - invalid_client_block_id
        - unsafe_content
        - duplicate_title
        - empty_course
        - unsupported_currency
        - invalid_input
        - exam_not_open
        - exam_closed
        - open_attempt_exists
        - retakes_not_allowed
        - attempt_submitted
        - attempt_expired
        - reveal_attempt_open
        - reveal_disabled
        - reveal_not_yet
        - asset_in_use
        - invalid_quiz_reference
        - invalid_media_reference
        - payment_required
        - course_free
        - already_enrolled
        - already_purchased
        - payment_unavailable
    QuizOption:
      type: object
      required: [id, label]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        label: { type: string }
    QuizQuestion:
      type: object
      required: [id, prompt, type, options, correct_option_ids, explanation]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        prompt: { type: string }
        type: { type: string, enum: [single_choice, multiple_choice, true_false] }
        options:
          type: array
          items: { $ref: "#/components/schemas/QuizOption" }
        correct_option_ids:
          type: array
          items: { $ref: "#/components/schemas/ID" }
        explanation: { type: string }
    Quiz:
      type: object
      required: [id, lecture_id, position, revision, questions]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        lecture_id: { $ref: "#/components/schemas/ID" }
        position: { type: integer, minimum: 0 }
        revision: { type: integer, minimum: 1 }
        questions:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestion" }
    SaveQuizRequest:
      type: object
      required: [position, questions]
      additionalProperties: false
      properties:
        position: { type: integer, minimum: 0, maximum: 10000 }
        questions:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [prompt, type, options]
            additionalProperties: false
            properties:
              id: { type: string, description: Empty or omitted for a new question }
              prompt: { type: string, maxLength: 5000 }
              type: { type: string, enum: [single_choice, multiple_choice, true_false] }
              explanation: { type: string, maxLength: 5000 }
              points: { type: integer, minimum: 1, description: Defaults to 1 }
              reference_lecture_id: { type: string, description: Empty or omitted when absent }
              options:
                type: array
                minItems: 2
                maxItems: 10
                items:
                  type: object
                  required: [label, is_correct]
                  additionalProperties: false
                  properties:
                    id: { type: string, description: Empty or omitted for a new option }
                    label: { type: string, maxLength: 1000 }
                    is_correct: { type: boolean }
    RecordQuizAttemptRequest:
      type: object
      required: [user_id, answers]
      additionalProperties: false
      properties:
        user_id: { $ref: "#/components/schemas/ID" }
        answers:
          type: array
          items:
            type: object
            required: [question_id, option_ids]
            additionalProperties: false
            properties:
              question_id: { $ref: "#/components/schemas/ID" }
              option_ids:
                type: array
                items: { $ref: "#/components/schemas/ID" }
    QuizAttemptResult:
      type: object
      required: [id, recorded]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        recorded: { type: boolean }
...